move docker builder struct into correct module
